fixed and refactored auth jest test
