+++ app/addons/fauxton/navigation/tests/storeSpec.js
@@ -9,7 +9,6 @@
 // WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 // License for the specific language governing permissions and limitations under
 // the License.
-import app from "../../../../app";
 import testUtils from "../../../../../test/mocha/testUtils";
 import FauxtonAPI from "../../../../core/api";
 import Stores from "../stores";
